+++ app/controllers/carto/api/visualizations_controller.rb
@@ -94,11 +94,17 @@ module Carto
         visualizations = vqb.with_order(order, order_direction)
                             .build_paged(page, per_page).map do |v|
           VisualizationPresenter.new(v, current_viewer, self, presenter_options)
-                                .with_presenter_cache(presenter_cache).to_poro
-        end
+                                .with_presenter_cache(presenter_cache).to_poro \
+            unless params[:subscribed] == 'true' and not v.subscription.present?
+        end.compact
+
+        total_subscriptions = 0
+        vqb.filtered_query.find_each{|v| total_subscriptions += 1 if v.subscription.present?}
+
         response = {
           visualizations: visualizations,
-          total_entries: vqb.count
+          total_entries: vqb.count,
+          total_subscriptions: total_subscriptions
         }
         if current_user && (params[:load_totals].to_s != 'false')
           response.merge!(calculate_totals(types))