+++ app/models/carto/visualization.rb
@@ -658,6 +658,14 @@ class Carto::Visualization < ActiveRecord::Base
     end
   end
 
+  def subscription
+    if user_table
+      doss = Carto::DoSyncService.new(user)
+      subscription_id = doss.subscription_from_sync_table(user_table.name)
+      { type: 'do-v2', id: subscription_id } if subscription_id.present?
+    end
+  end
+
   private
 
   def remove_password